refector: move rooms.ts to RoomsService
